MQTTServer/requirements.txt
paho-mqtt
Flask
Flask-SocketIO